backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "irli-backend"
version = "0.1.0"
description = "IRLI — Israel Research Lab Index extraction pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic[email]>=2.7.0",
    "python-dotenv>=1.0.1",
    "crawl4ai>=0.3.0",
    "openai>=1.30.0",
    "instructor>=1.3.0",
    "google-generativeai>=0.7.0",
    "httpx>=0.27.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "pgvector>=0.3",
    "apscheduler>=3.10",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
